+++ modules/src/ics03_connection/msgs.rs
@@ -22,15 +22,6 @@ pub mod conn_open_confirm;
 pub mod conn_open_init;
 pub mod conn_open_try;
 
-/// Enumeration of all possible message types that the ICS3 protocol processes.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum ConnectionMsgType {
-    OpenInit,
-    OpenTry,
-    OpenAck,
-    OpenConfirm,
-}
-
 /// Enumeration of all possible messages that the ICS3 protocol processes.
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum ConnectionMsg {