+++ modules/src/ics04_channel/channel.rs
@@ -2,11 +2,13 @@ use crate::ics04_channel::error::{self, Error, Kind};
 use crate::ics24_host::identifier::{ChannelId, ConnectionId, PortId};
 
 use ibc_proto::ibc::core::channel::v1::Channel as RawChannel;
+use ibc_proto::ibc::core::channel::v1::Counterparty as RawCounterparty;
+
+use tendermint_proto::Protobuf;
 
 use anomaly::fail;
-use std::convert::TryFrom;
+use std::convert::{TryFrom, TryInto};
 use std::str::FromStr;
-use tendermint_proto::DomainType;
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct ChannelEnd {
@@ -17,7 +19,7 @@ pub struct ChannelEnd {
     version: String,
 }
 
-impl DomainType<RawChannel> for ChannelEnd {}
+impl Protobuf<RawChannel> for ChannelEnd {}
 
 impl TryFrom<RawChannel> for ChannelEnd {
     type Error = anomaly::Error<Kind>;
@@ -28,19 +30,11 @@ impl TryFrom<RawChannel> for ChannelEnd {
 
         let chan_state = State::from_i32(value.state)?;
 
-        // Pop out the Counterparty from the Option.
-        let counterparty = match value.counterparty {
-            Some(cp) => cp,
-            None => return Err(Kind::MissingCounterparty.into()),
-        };
-
         // Assemble the 'remote' attribute of the Channel, which represents the Counterparty.
-        let remote = Counterparty {
-            port_id: PortId::from_str(counterparty.port_id.as_str())
-                .map_err(|e| Kind::IdentifierError.context(e))?,
-            channel_id: ChannelId::from_str(counterparty.channel_id.as_str())
-                .map_err(|e| Kind::IdentifierError.context(e))?,
-        };
+        let remote = value
+            .counterparty
+            .ok_or(Kind::MissingCounterparty)?
+            .try_into()?;
 
         // Parse each item in connection_hops into a ConnectionId.
         let connection_hops = value
@@ -50,14 +44,15 @@ impl TryFrom<RawChannel> for ChannelEnd {
             .collect::<Result<Vec<_>, _>>()
             .map_err(|e| Kind::IdentifierError.context(e))?;
 
-        // This field is supposed to be opaque to the core IBC protocol. Empty
-        // version is allowed by the specification (cf. ICS 004). No explicit validation necessary.
-        let version = value.version;
+        let version = validate_version(value.version)?;
 
-        let mut channel_end = ChannelEnd::new(chan_ordering, remote, connection_hops, version);
-        channel_end.set_state(chan_state);
-
-        Ok(channel_end)
+        Ok(ChannelEnd::new(
+            chan_state,
+            chan_ordering,
+            remote,
+            connection_hops,
+            version,
+        ))
     }
 }
 
@@ -65,11 +60,8 @@ impl From<ChannelEnd> for RawChannel {
     fn from(value: ChannelEnd) -> Self {
         RawChannel {
             state: value.state.clone() as i32,
-            ordering: value.ordering.clone() as i32,
-            counterparty: Some(ibc_proto::ibc::core::channel::v1::Counterparty {
-                port_id: value.counterparty().port_id.to_string(),
-                channel_id: value.counterparty().channel_id.to_string(),
-            }),
+            ordering: value.ordering as i32,
+            counterparty: Some(value.counterparty().into()),
             connection_hops: value
                 .connection_hops
                 .iter()
@@ -83,13 +75,14 @@ impl From<ChannelEnd> for RawChannel {
 impl ChannelEnd {
     /// Creates a new ChannelEnd in state Uninitialized and other fields parametrized.
     pub fn new(
+        state: State,
         ordering: Order,
         remote: Counterparty,
         connection_hops: Vec<ConnectionId>,
         version: String,
     ) -> Self {
         Self {
-            state: State::Uninitialized,
+            state,
             ordering,
             remote,
             connection_hops,
@@ -137,28 +130,24 @@ impl ChannelEnd {
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct Counterparty {
-    port_id: PortId,
-    channel_id: ChannelId,
+    pub port_id: PortId,
+    pub channel_id: Option<ChannelId>,
 }
 
 impl Counterparty {
-    pub fn new(port_id: String, channel_id: String) -> Result<Self, Error> {
-        Ok(Self {
-            port_id: port_id
-                .parse()
-                .map_err(|e| Kind::IdentifierError.context(e))?,
-            channel_id: channel_id
-                .parse()
-                .map_err(|e| Kind::IdentifierError.context(e))?,
-        })
+    pub fn new(port_id: PortId, channel_id: Option<ChannelId>) -> Self {
+        Self {
+            port_id,
+            channel_id,
+        }
     }
 
-    pub fn port_id(&self) -> String {
-        self.port_id.as_str().into()
+    pub fn port_id(&self) -> &PortId {
+        &self.port_id
     }
 
-    pub fn channel_id(&self) -> String {
-        self.channel_id.as_str().into()
+    pub fn channel_id(&self) -> Option<&ChannelId> {
+        self.channel_id.as_ref()
     }
 
     pub fn validate_basic(&self) -> Result<(), Error> {
@@ -166,11 +155,49 @@ impl Counterparty {
     }
 }
 
-#[derive(Clone, Debug, PartialEq)]
+impl Protobuf<RawCounterparty> for Counterparty {}
+
+impl TryFrom<RawCounterparty> for Counterparty {
+    type Error = anomaly::Error<Kind>;
+
+    fn try_from(value: RawCounterparty) -> Result<Self, Self::Error> {
+        let channel_id = Some(value.channel_id)
+            .filter(|x| !x.is_empty())
+            .map(|v| FromStr::from_str(v.as_str()))
+            .transpose()
+            .map_err(|e| Kind::IdentifierError.context(e))?;
+        Ok(Counterparty::new(
+            value
+                .port_id
+                .parse()
+                .map_err(|e| Kind::IdentifierError.context(e))?,
+            channel_id,
+        ))
+    }
+}
+
+impl From<Counterparty> for RawCounterparty {
+    fn from(value: Counterparty) -> Self {
+        RawCounterparty {
+            port_id: value.port_id.as_str().to_string(),
+            channel_id: value
+                .channel_id
+                .map_or_else(|| "".to_string(), |v| v.as_str().to_string()),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub enum Order {
     None = 0,
     Unordered,
     Ordered,
+}
+
+impl Default for Order {
+    fn default() -> Self {
+        Order::Unordered
+    }
 }
 
 impl Order {
@@ -241,37 +268,55 @@ impl State {
     }
 }
 
+/// Version validation, specific for channel (ICS4) opening handshake protocol.
+/// This field is supposed to be opaque to the core IBC protocol. No explicit validation necessary,
+/// and empty version is currently allowed by the specification (cf. ICS 004, v1).
+pub fn validate_version(version: String) -> Result<String, Error> {
+    Ok(version)
+}
+
+#[cfg(test)]
+pub mod test_util {
+    use ibc_proto::ibc::core::channel::v1::Channel as RawChannel;
+    use ibc_proto::ibc::core::channel::v1::Counterparty as RawCounterparty;
+
+    /// Returns a dummy `RawCounterparty`, for testing only!
+    pub fn get_dummy_raw_counterparty() -> RawCounterparty {
+        RawCounterparty {
+            port_id: "0123456789".into(),
+            channel_id: "0987654321".into(),
+        }
+    }
+
+    /// Returns a dummy `RawChannel`, for testing only!
+    pub fn get_dummy_raw_channel_end() -> RawChannel {
+        RawChannel {
+            state: 0,
+            ordering: 0,
+            counterparty: Some(get_dummy_raw_counterparty()),
+            connection_hops: vec![],
+            version: "".to_string(), // The version is not validated.
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use std::str::FromStr;
 
+    use crate::ics04_channel::channel::test_util::get_dummy_raw_channel_end;
     use crate::ics04_channel::channel::ChannelEnd;
 
     use ibc_proto::ibc::core::channel::v1::Channel as RawChannel;
-    use ibc_proto::ibc::core::channel::v1::Counterparty as RawCounterparty;
     use std::convert::TryFrom;
 
     #[test]
     fn channel_end_try_from_raw() {
+        let raw_channel_end = get_dummy_raw_channel_end();
+
         let empty_raw_channel_end = RawChannel {
-            state: 0,
-            ordering: 0,
             counterparty: None,
-            connection_hops: vec![],
-            version: "".to_string(),
-        };
-
-        let cparty = RawCounterparty {
-            port_id: "0123456789".into(),
-            channel_id: "0987654321".into(),
-        };
-
-        let raw_channel_end = RawChannel {
-            state: 0,
-            ordering: 0,
-            counterparty: Some(cparty),
-            connection_hops: vec![],
-            version: "".to_string(), // The version is not validated.
+            ..raw_channel_end.clone()
         };
 
         struct Test {