+++ relayer/src/chain/handle.rs
@@ -1,15 +1,22 @@
-use std::time::Duration;
-
-use crossbeam_channel as channel;
-use tendermint::block::signed_header::SignedHeader;
-use thiserror::Error;
-
-use ibc::{ics24_host::identifier::ChainId, Height};
-
+use crate::config::ChainConfig;
 use crate::foreign_client::ForeignClient;
 use crate::msgs::{Datagram, EncodedTransaction, IBCEvent, Packet};
+use crate::util::block_on;
 
-pub type Datagrams = Vec<Datagram>;
+use ibc::ics24_host::{identifier::ChainId, Path, IBC_QUERY_PATH};
+use ibc::Height;
+
+use tendermint::abci::Path as ABCIPath;
+use tendermint::net;
+use tendermint_rpc::HttpClient;
+
+use crossbeam_channel as channel;
+use ibc::ics02_client::client_def::{AnyClientState, AnyConsensusState, AnyHeader};
+use std::str::FromStr;
+use std::time::Duration;
+use thiserror::Error;
+
+pub(crate) mod cosmos; // Implementation of handle specific for Cosmos SDK chains.
 
 // Simplified:
 // Subscriptions should have provide processing semantics such
@@ -24,149 +31,52 @@ pub type Subscription = channel::Receiver<(Height, Vec<IBCEvent>)>;
 pub enum ChainHandleError {
     #[error("failed")]
     Failed,
+
+    #[error("requested proof for data in the privateStore")]
+    NonProvableData,
+
+    #[error("rpc client returned error: {0}")]
+    RPC(String),
+
+    #[error("invalid chain identifier format: {0}")]
+    ChainIdentifier(String),
+
+    #[error("the input header is not recognized as a header for this chain")]
+    InvalidInputHeader,
 }
 
-pub trait ChainHandle: Send {
-    fn subscribe(&self, chain_id: ChainId) -> Result<Subscription, ChainHandleError>;
-
-    // Inclusion proofs
-    // It might be good to include an inclusion proof method which abstracts over the light client
-    // to prove that a peice of data is stored on the chain
-
-    // TODO: Error Handling, get rid of this?
-    fn get_header(&self, height: Height) -> SignedHeader;
-
-    fn get_minimal_set(
-        &self,
-        from: Height,
-        to: Height,
-    ) -> Result<Vec<SignedHeader>, ChainHandleError>;
-
-    fn submit(&self, transaction: EncodedTransaction) -> Result<(), ChainHandleError>;
-
-    // Mocked:
-    // - query the consensus_state of src on dst
-    // - query the highest consensus_state
-    // - verify if with the light client
-    // - return the height
-    // + TODO: Can eventually be populated be pre-populated by a event_monitor subscription to the
-    // to the full node
-    fn get_height(&self, client: &ForeignClient) -> Result<Height, ChainHandleError>;
-
-    fn id(&self) -> ChainId;
-
-    fn create_packet(&self, event: IBCEvent) -> Result<Packet, ChainHandleError>;
-}
-
-#[derive(Debug, Clone)]
-pub struct ProdChainHandle {
-    pub chain_id: ChainId,
-    sender: channel::Sender<HandleInput>,
-    // TODO: account_prefix
-}
-
-impl ProdChainHandle {
-    fn new(chain_id: ChainId, sender: channel::Sender<HandleInput>) -> Self {
-        Self { chain_id, sender }
-    }
-}
-
-impl ChainHandle for ProdChainHandle {
-    fn id(&self) -> ChainId {
-        self.chain_id.clone()
-    }
-
-    fn subscribe(&self, _chain_id: ChainId) -> Result<Subscription, ChainHandleError> {
-        let (sender, receiver) = channel::bounded::<Subscription>(1);
-        self.sender.send(HandleInput::Subscribe(sender)).unwrap();
-        Ok(receiver.recv().unwrap())
-    }
-
-    fn get_header(&self, height: Height) -> SignedHeader {
-        todo!()
-    }
-
-    fn get_minimal_set(
-        &self,
-        from: Height,
-        to: Height,
-    ) -> Result<Vec<SignedHeader>, ChainHandleError> {
-        todo!()
-    }
-
-    fn submit(&self, _transaction: EncodedTransaction) -> Result<(), ChainHandleError> {
-        todo!()
-    }
-
-    fn get_height(&self, _client: &ForeignClient) -> Result<Height, ChainHandleError> {
-        todo!()
-    }
-
-    fn create_packet(&self, _event: IBCEvent) -> Result<Packet, ChainHandleError> {
-        todo!()
-    }
-}
-
+// Inputs that a Handle may send to a Runtime.
 pub enum HandleInput {
     Terminate(channel::Sender<()>),
     Subscribe(channel::Sender<Subscription>),
 }
 
-pub struct ChainRuntime {
-    chain_id: ChainId,
-    sender: channel::Sender<HandleInput>,
-    receiver: channel::Receiver<HandleInput>,
-}
+pub trait ChainHandle: Send {
+    fn subscribe(&self, chain_id: ChainId) -> Result<Subscription, ChainHandleError>;
 
-impl ChainRuntime {
-    // XXX: ChainConfig
-    pub fn new(chain_id: ChainId) -> ChainRuntime {
-        let (sender, receiver) = channel::unbounded::<HandleInput>();
+    fn query(&self, path: Path, height: Height, prove: bool) -> Result<Vec<u8>, ChainHandleError>;
 
-        Self {
-            chain_id,
-            sender,
-            receiver,
-        }
-    }
+    fn get_header(&self, height: Height) -> Result<AnyHeader, ChainHandleError>;
 
-    pub fn handle(&self) -> ProdChainHandle {
-        let sender = self.sender.clone();
-        ProdChainHandle::new(self.chain_id.clone(), sender)
-    }
+    fn get_minimal_set(&self, from: Height, to: Height)
+        -> Result<Vec<AnyHeader>, ChainHandleError>;
 
-    pub fn run(self) -> Result<(), ChainHandleError> {
-        // Mocked: EventMonitor
-        // What we need here is a reliable stream of events produced by a connected full node.
-        // Events received from this stream will be buffered (perhaps durably) and then routed to
-        // the various subscriptions.
-        let event_monitor = channel::tick(Duration::from_millis(1000));
+    /// Submits a transaction.
+    fn submit(&self, transaction: EncodedTransaction) -> Result<(), ChainHandleError>;
 
-        let mut subscriptions: Vec<channel::Sender<(Height, Vec<IBCEvent>)>> = vec![];
-        loop {
-            channel::select! {
-                recv(event_monitor) -> tick => {
-                    println!("tick tick!");
-                    for subscription in subscriptions.iter() {
-                        let target_height = Height::new(0, 1); // FIXME
-                        // subscription.send((target_height, vec![IBCEvent::NoOp()])).unwrap();
-                    }
-                },
-                recv(self.receiver) -> maybe_event => {
-                    let event = maybe_event.unwrap();
-                    match event {
-                        HandleInput::Subscribe(sender) => {
-                            println!("Subscribing!");
-                            let (sub_sender, sub_receiver) = channel::unbounded::<(Height, Vec<IBCEvent>)>();
-                            subscriptions.push(sub_sender);
-                            sender.send(sub_receiver).unwrap();
-                        },
-                        HandleInput::Terminate(sender) => {
-                            sender.send(()).unwrap();
-                        }
-                    }
-                },
-            }
-        }
-    }
+    fn get_height(&self, client: &ForeignClient) -> Result<Height, ChainHandleError>;
+
+    fn id(&self) -> ChainId;
+
+    fn create_packet(&self, event: IBCEvent) -> Result<Packet, ChainHandleError>;
+
+    /// Given a header originating from this chain, constructs a client state.
+    fn assemble_client_state(&self, header: &AnyHeader)
+        -> Result<AnyClientState, ChainHandleError>;
+
+    /// Given a header originating from this chain, constructs a consensus state.
+    fn assemble_consensus_state(
+        &self,
+        header: &AnyHeader,
+    ) -> Result<AnyConsensusState, ChainHandleError>;
 }